requirements.txt
asgiref==3.6.0
boto3==1.26.88
botocore==1.29.88
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
